homemanager_backend/requirements.txt
# Requirements for HomeManager backend
Django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=24.1
psycopg2-binary>=2.9.9
Pillow>=10.1.0
python-dateutil>=2.8.2
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
django-storages>=1.14.2
requests>=2.31.0
pytz>=2023.3
uuid>=1.30
python-decouple>=3.8
django-debug-toolbar>=4.2.0
django-environ>=0.11.2
drf-yasg>=1.21.7
sentry-sdk>=1.39.1
celery>=5.3.6
redis>=5.0.1
python-magic>=0.4.27
django-crispy-forms>=2.1
faker>=22.0.0
